Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -6,7 +6,7 @@ int main() {
     float r;
     scanf("%f",r);
     float pie = 3.14;
-    double area = pie*r*r;
+    double area =(float)(pie*r*r);
     printf("Area: %.2lf",area);
     return 0;
 }
